+++ mininote-frontend/src/pages/App/components/SettingsModal.vue
@@ -18,8 +18,8 @@
 </template>
 
 <script>
-  import NotesApiService from "../../../services/NotesApiService";
-  import { md5 } from "../../../services/md5";
+  import NotesApiService from "../services/NotesApiService";
+  import { md5 } from "../services/md5";
 
   export default {
     name: 'settingsmodal',
