+++ mininote-frontend/src/pages/App/pages/NotebookListPage/NotebookListPage.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-  import NotesApiService from "../../../../services/NotesApiService";
+  import NotesApiService from "../../services/NotesApiService";
 
   import NotebookList from './components/NotebookList'
