+++ mininote-frontend/src/pages/App/services/marked.js
