+++ mininote-frontend/src/pages/App/services/md5.js
